trim ./application test files
